chore: Update SuperLinter EditorConfig Checker (#195)
# Pull Request

## Description

This pull request includes a couple of changes to improve the
configuration and workflow files in the repository.

Configuration updates:

*
[`.github/super-linter-configs/.editorconfig-checker.json`](diffhunk://#diff-35faced47eceeb79bd71f718bc0399c19709d75eb137af5a164e97d17346666aL5-R5):
Corrected a typo in the `SpacesAfterTabs` key and renamed the file from
`.ecrc` to `.editorconfig-checker.json`.

Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R35):
Added the `EDITORCONFIG_FILE_NAME` environment variable to specify the
renamed `.editorconfig-checker.json` file.