ci: Add Lefthook Validate Job (#202)
# Pull Request

## Description

This pull request introduces a new job to the GitHub Actions workflow
configuration for validating Lefthook hooks. The addition ensures that
Lefthook validation is integrated into the CI pipeline.

### Workflow Enhancements:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R181-R196):
Added a new job named `lefthook-validate` to the workflow. This job runs
on `ubuntu-latest`, checks out the repository, installs the latest
version of `uv`, and executes the `lefthook validate` command to
validate Lefthook hooks.