chore: Fix Zizmor SARIF Command (#222)
# Pull Request

## Description

This pull request updates the workflow and build configuration to
improve the integration of tools and streamline the `zizmor` checks. The
most important changes include upgrading the `setup-uv` action,
introducing the `setup-just` action, and modifying commands for `zizmor`
checks to use `Just` for better consistency.

### Workflow updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L157-R165):
Upgraded the `setup-uv` action to version `v6.0.1` and added the
`setup-just` action (v3) to set up `Just`, a command runner. Updated the
`zizmor` command to use `just zizmor-check-sarif` for running checks and
generating SARIF output. Additionally, updated the `upload-sarif` action
to version `v3.28.17`.

### Build configuration updates:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cL83-R87):
Added a new `zizmor-check-sarif` command that uses `uvx` to run `zizmor`
with SARIF output. Replaced the previous `zizmor-check` command with an
updated version that uses `uvx` for consistency.