chore: Bump Go Version (#247)
# Pull Request

## Description

This pull request includes updates to the Go version to address
potential security or stability improvements. The changes ensure
consistency across the project by updating the base image in the
`Dockerfile` and the Go version in `go.mod`.

Version updates:

*
[`Dockerfile`](diffhunk://#diff-dd2c0eb6ea5cfc6c4bd4eac30934e2d5746747af48fef6da689e85b752f39557L3-R3):
Updated the base image from `golang:1.24.0` to `golang:1.24.3`.
*
[`go.mod`](diffhunk://#diff-33ef32bf6c23acb95f5902d7097b7a1d5128ca061167ec0716715b0b9eeaa5f6L3-R3):
Updated the Go version from `1.24.0` to `1.24.3`.