deps(go): Bump go-github to v74 (#325)
# Pull Request

## Description

This pull request updates the dependency on the `go-github` library to a
newer major version throughout the codebase. This change ensures
compatibility with the latest features and bug fixes from the library.

Dependency upgrade:

* Updated `github.com/google/go-github` from version `v68` to `v74` in
both `go.mod` and the import statement in `main.go`, ensuring the
project uses the latest API and improvements.
[[1]](diffhunk://#diff-33ef32bf6c23acb95f5902d7097b7a1d5128ca061167ec0716715b0b9eeaa5f6L6-R6)
[[2]](diffhunk://#diff-2873f79a86c0d8b3335cd7731b0ecf7dd4301eb19a82ef7a1cba7589b5252261L9-R9)