ci: Update Pull Request Tasks Permissions (#357)
# Pull Request

## Description

This pull request makes a minor update to the GitHub Actions workflow
configuration by relocating the `pull-requests: read` permission from
the top-level `permissions` block to the job-specific `permissions`
block for the `check-pull-request-title` job. This change ensures that
only the relevant job is granted the necessary permissions, improving
the principle of least privilege.

- Workflow permissions update:
* Moved the `pull-requests: read` permission from the global workflow
level to the `check-pull-request-title` job level in
`.github/workflows/pull-request-tasks.yml`, setting the top-level
permissions to an empty object.